Add message_tracker
