+++ bot/handlers/on_message.py
@@ -6,7 +6,7 @@ from discord.ui import View
 from bot.config import config
 from bot.handlers.record_mission_handler import handle_record_mission_start, handle_record_mission_dm
 from bot.handlers.video_mission_handler import handle_video_mission_start, handle_video_mission_dm
-from bot.views.optin_class import OptinClassView
+from bot.handlers.utils import handle_greeting_job
 
 async def handle_dm(client, message):
     if (
@@ -47,44 +47,6 @@ async def handle_dm(client, message):
             else:
                 await handle_video_mission_dm(client, message, student_mission_info)
         return
-
-async def handle_greeting_job(client, user_id = None):
-    hello_message = (
-        "🐾 欸～新手爸媽們！我是加一，你的「寶寶照護教室」導師！\n\n"
-        "照顧寶寶是不是覺得像進入新手村？\n"
-        "別怕，有我罩你！💪 交給我，穩穩的！😆 \n"
-        "奶瓶怎麼選？尿布怎麼換？寶寶半夜哭鬧怎麼辦？\n"
-        "專屬課程手把手帶你\n"
-        "讓你穩穩當當晉升帶娃高手！🍼\n\n"
-        "📣 有問題？盡管問！ 課堂直接解答，別再半夜上網查到懷疑人生～📲\n"
-        "新手爸媽，不用怕，你肯定行！ 加一帶你穩穩走～💪\n"
-        "📌 快來看看課程重點，直接登記加入！ 🌟\n"
-        "四個月大以上的寶寶也可以登記喔！"
-    )
-
-    if user_id == None:
-        student_list = await client.api_utils.fetch_student_list()
-    else:
-        student_list = [{'discord_id': user_id}]
-
-    # start greeting
-    client.logger.info(f"Start greeting job: {len(student_list)} student")
-    for user in student_list:
-        files = [
-            discord.File("bot/resource/mission_bot_1.png"),
-            discord.File("bot/resource/mission_bot_2.png"),
-            discord.File("bot/resource/mission_bot_3.png"),
-            discord.File("bot/resource/mission_bot_4.png")
-        ]
-        user_id = user['discord_id']
-        user = await client.fetch_user(user_id)
-        view = OptinClassView(client, user_id)
-        view.message = await user.send(hello_message, view=view, files=files)
-        client.logger.info(f"Send hello message to user {user_id}")
-        await client.api_utils.store_message(user_id, 'assistant', hello_message)
-        await asyncio.sleep(10)
-
-    return
 
 async def handle_start_mission(client, user_id, mission_id):
     mission_id = int(mission_id)
